requirements.txt
djangorestframework
djangorestframework-simplejwt
dj-rest-auth
django-allauth
drf-spectacular
django-cors-headers
argon2-cffi
